+++ modules/kubernetes/aks-core/modules.tf
@@ -18,7 +18,6 @@ locals {
     "starboard-operator",
     "tigera-operator",
     "velero",
-    "newrelic",
     "grafana-agent",
   ]
 }
@@ -487,20 +486,6 @@ module "vpa" {
   }
 
   source = "../../kubernetes/vpa"
-}
-
-module "new_relic" {
-  for_each = {
-    for s in ["new-relic"] :
-    s => s
-    if var.new_relic_enabled
-  }
-
-  source = "../../kubernetes/new-relic"
-
-  cluster_name      = "${var.name}${var.aks_name_suffix}-${var.environment}-${var.location_short}"
-  license_key       = var.new_relic_config.license_key
-  namespace_include = var.namespaces[*].name
 }
 
 # node-local-dns