+++ modules/kubernetes/aks-core/variables.tf
@@ -417,22 +417,6 @@ variable "starboard_config" {
   })
 }
 
-variable "new_relic_enabled" {
-  description = "Should New Relic be enabled"
-  type        = bool
-  default     = false
-}
-
-variable "new_relic_config" {
-  description = "Configuration for New Relic"
-  type = object({
-    license_key = string
-  })
-  default = {
-    license_key = ""
-  }
-}
-
 variable "azure_metrics_enabled" {
   description = "Should AZ Metrics be enabled"
   type        = bool