+++ modules/kubernetes/eks-core/variables.tf
@@ -313,22 +313,6 @@ variable "datadog_config" {
   }
 }
 
-variable "new_relic_enabled" {
-  description = "Should New Relic be enabled"
-  type        = bool
-  default     = false
-}
-
-variable "new_relic_config" {
-  description = "Configuration for New Relic"
-  type = object({
-    license_key = string
-  })
-  default = {
-    license_key = ""
-  }
-}
-
 variable "starboard_enabled" {
   description = "Should Starboard be enabled"
   type        = bool